requirements.txt
Flask==2.2.5
flask-cors==3.0.10
Pillow==9.4.0
google-generativeai==1.8.4
gunicorn